agent-os-runtime/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-os-runtime"
version = "0.6.0"
description = "Agent OS Runtime: Agno + Mem0 + Hindsight + Graphiti + doctor/retry"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "agent-os-runtime contributors" }]
dependencies = [
  "agno>=1.0",
  "mem0ai>=0.1",
  "pydantic>=2.0",
  "python-dotenv>=1.0.0",
  "openai>=1.40.0",
]

[project.optional-dependencies]
graphiti = [
  "graphiti-core>=0.28.0,<1",
]
mcp = [
  "mcp>=1.0.0",
]
asset_store = [
  "lancedb",
]
dev = ["ruff>=0.4.0", "pytest>=8.0.0", "lancedb", "pre-commit>=3.5"]
web = ["fastapi>=0.115.0", "uvicorn[standard]>=0.32.0"]

[project.scripts]
agent-os-runtime = "agent_os.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agent_os = ["py.typed", "resources/*.json", "data/skill_manifests/*.json"]

[tool.ruff]
target-version = "py310"
src = ["src", "tests"]
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
markers = [
  "skill_example: 示例 skill 回归",
]
